# Restore configure_recurse; use local reconfigure for dynamic parents