--- a/crates/kas-widgets/src/grid.rs
+++ b/crates/kas-widgets/src/grid.rs
@@ -139,10 +139,15 @@ impl<W: Widget> Grid<W> {
     }
 
     /// Edit an existing grid via a builder
+    ///
+    /// This may be used to edit children before window construction. It may
+    /// also be used from a running UI, but in this case a full reconfigure
+    /// of the window's widgets is required (triggered by the the return
+    /// value, [`TkAction::RECONFIGURE`]).
     pub fn edit<F: FnOnce(GridBuilder<W>)>(&mut self, f: F) -> TkAction {
         f(GridBuilder(&mut self.widgets));
         self.calc_dim();
-        TkAction::RECONFIGURE // just assume this is requried
+        TkAction::RECONFIGURE
     }
 
     /// True if there are no child widgets