--- a/crates/kas-widgets/src/list.rs
+++ b/crates/kas-widgets/src/list.rs
@@ -259,6 +259,18 @@ widget! {
             }
         }
 
+        /// Edit the list of children directly
+        ///
+        /// This may be used to edit children before window construction. It may
+        /// also be used from a running UI, but in this case a full reconfigure
+        /// of the window's widgets is required (triggered by the the return
+        /// value, [`TkAction::RECONFIGURE`]).
+        #[inline]
+        pub fn edit<F: FnOnce(&mut Vec<W>)>(&mut self, f: F) -> TkAction {
+            f(&mut self.widgets);
+            TkAction::RECONFIGURE
+        }
+
         /// Get the direction of contents
         pub fn direction(&self) -> Direction {
             self.direction.as_direction()