--- a/crates/kas-widgets/src/splitter.rs
+++ b/crates/kas-widgets/src/splitter.rs
@@ -6,7 +6,7 @@
 //! A row or column with sizes adjustable via dividing handles
 
 use log::warn;
-use std::collections::HashMap;
+use std::collections::hash_map::{Entry, HashMap};
 use std::ops::{Index, IndexMut};
 
 use super::DragHandle;
@@ -105,8 +105,8 @@ widget! {
             loop {
                 let key = self.next;
                 self.next += 1;
-                if !self.id_map.contains_key(&key) {
-                    self.id_map.insert(key, child_index);
+                if let Entry::Vacant(entry) = self.id_map.entry(key) {
+                    entry.insert(index);
                     return self.id_ref().make_child(key);
                 }
             }