--- a/crates/kas-widgets/src/stack.rs
+++ b/crates/kas-widgets/src/stack.rs
@@ -6,7 +6,7 @@
 //! A stack
 
 use kas::{event, prelude::*};
-use std::collections::HashMap;
+use std::collections::hash_map::{Entry, HashMap};
 use std::fmt::Debug;
 use std::ops::{Index, IndexMut, Range};
 
@@ -63,8 +63,8 @@ widget! {
             loop {
                 let key = self.next;
                 self.next += 1;
-                if !self.id_map.contains_key(&key) {
-                    self.id_map.insert(key, index);
+                if let Entry::Vacant(entry) = self.id_map.entry(key) {
+                    entry.insert(index);
                     return self.id_ref().make_child(key);
                 }
             }