… crates/kas-widgets/src/view/matrix_view.rs
… crates/kas-widgets/src/view/single_view.rs
… examples/data-list.rs
… examples/gallery.rs
… examples/splitter.rs