`List`, `Stack`, `ComboBox` and `Splitter` widgets now locally reconfigure new children without changing the `WidgetId` of other children (they use a `HashMap` for `WidgetId` → `child_index` lookup).

The result (using `example/data-list.rs` and a hand-controlled stopwatch):

| | before | after |
|- | -| -|
| from 0 to 100,000 | 17.6s | 17.8s |
| from 100,000 to 100,001 | 13.8s | 0.6s |

Full-window resize is still required after adjusting the number of widgets, but apparently this is clearly much faster than configuring widgets (a little surprising since configuring is mostly just assigning a `WidgetId`).

TODO: `ComboBox` doesn't need to resize the window if the menu is closed, and only needs to resize the menu if open, but it inherits the full-window resize from `List`. (Low impact optimisation so I'll leave it.)